pyproject.toml
[project]
name = "ada"
version = "0.1.0"
description = "Artifact-driven agent — tools generated from markdown specs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "loguru>=0.7",
    "pydantic>=2.0",
    "pydantic-ai>=1.80.0",
    "pydantic-monty>=0.0.11",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ada"]
